move common files to common directory
